Explicit long to int conversions  (for 76172).
git-svn-id: https://svn.r-project.org/R/trunk@76183 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/character.c
@@ -301,7 +301,7 @@ static void substr(const char *str, int len, int ienc, int sa, int so,
 	*rfrom = str;
 	for(; i < so && str < end; i++)
 	    str += utf8clen(*str);
-	*rlen = str - *rfrom;
+	*rlen = (int) (str - *rfrom);
     } else if (!isascii && ienc != CE_LATIN1 && ienc != CE_BYTES
                && mbcslocale) {
 	mbstate_t mb_st;
@@ -313,7 +313,7 @@ static void substr(const char *str, int len, int ienc, int sa, int so,
 	for (; i < so && str < end; i++)
 	    /* throws error on invalid multi-byte string */
 	    str += (int) Mbrtowc(NULL, str, MB_CUR_MAX, &mb_st);
-	*rlen = str - *rfrom;
+	*rlen = (int) (str - *rfrom);
     } else {
 	if (so - 1 < len) {
 	    *rfrom = str + sa - 1;
@@ -358,7 +358,7 @@ do_substr(SEXP call, SEXP op, SEXP args, SEXP env)
 	    }
 	    cetype_t ienc = getCharCE(el);
 	    const char *ss = CHAR(el);
-	    size_t slen = LENGTH(el);
+	    int slen = LENGTH(el);
 	    if (start < 1) start = 1;
 	    if (start > stop) {
 		SET_STRING_ELT(s, i, R_BlankString);
